# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Avia-back API",
        "version": "1.0.3"
    },
    "paths": {
...
        "/api/v1/admin/technical-systems": {
            "get": {
                "tags": [
                    "Технические системы"
                ],
                "summary": "Получить список всех технических систем",
                "description": "Для супер-администратора возвращает список иерархий всех технических систем, созданных в системе. Для администратора возвращает список иерархий только тех технических систем, которые доступны в рамках проекта для той организации к которой принадлежит администратор.",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "code": {
                                                "type": "string",
                                                "example": "Some code"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Some name"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Some description"
                                            },
                                            "parent_technical_system_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-09-15T01:52:11.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-09-15T01:52:11.000000Z"
                                            },
                                            "documents": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "code": {
                                                            "type": "string",
                                                            "example": "Some code"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Some name"
                                                        },
                                                        "type": {
                                                            "type": "integer",
                                                            "example": 0
                                                        },
                                                        "version": {
                                                            "type": "string",
                                                            "example": "Some version"
                                                        },
                                                        "file": {
                                                            "type": "string",
                                                            "example": "Some path to file"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2023-09-15T01:52:11.000000Z"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2023-09-15T01:52:11.000000Z"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "grandchildren_technical_systems": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "code": {
                                                            "type": "string",
                                                            "example": "Some code"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Some name"
                                                        },
                                                        "description": {
                                                            "type": "string",
                                                            "example": "Some description"
                                                        },
                                                        "parent_technical_system_id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2023-09-15T01:52:11.000000Z"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2023-09-15T01:52:11.000000Z"
                                                        },
                                                        "grandchildren_technical_systems": {
                                                            "type": "array",
                                                            "items": {}
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Технические системы"
                ],
                "summary": "Создание новой технической системы или объекта",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "code": {
                                                "type": "string",
                                                "example": "Some code"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Some name"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Some description"
                                            },
                                            "parent_technical_system_id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "Some code"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Some name"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Some description"
                                        },
                                        "parent_technical_system_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-09-15T01:52:11.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-09-15T01:52:11.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}